+++ include/JoSIM/Input.hpp
@@ -22,7 +22,7 @@ class Input {
   public:
   Netlist netlist;
   Transient transSim;
-  std::optional<double> globalTemp, neB;
+  std::optional<float> globalTemp, neB;
   std::vector<tokens_t> fileLines, controls;
   std::vector<std::string> relevantX;
   std::unordered_map<ParameterName, Parameter> parameters;
