+++ include/JoSIM/VoltageSource.hpp
@@ -26,7 +26,7 @@ namespace JoSIM {
 class VoltageSource : public BasicComponent {
   public:
   int sourceIndex_;
-  double pn1_ = 0.0, pn2_ = pn1_, pn3_ = pn2_, pn4_ = 0.0;
+  float pn1_ = 0.0, pn2_ = pn1_, pn3_ = pn2_, pn4_ = 0.0;
   VoltageSource(
     const std::pair<tokens_t, string_o> &s, const NodeConfig &ncon,
     const nodemap &nm, std::unordered_set<std::string> &lm,
